Move progressbar actions to a Trait
