+++ src/Parallel/ParallelWorker.php
@@ -2,14 +2,11 @@
 
 namespace HDSSolutions\Console\Parallel;
 
-use Closure;
-use HDSSolutions\Console\Parallel\Internals\Messages\ProgressBarActionMessage;
-use HDSSolutions\Console\Parallel\Internals\Messages\StatsReportMessage;
-use parallel\Channel;
 use RuntimeException;
 use Throwable;
 
 abstract class ParallelWorker implements Contracts\ParallelWorker {
+    use Internals\Worker\CommunicatesWithProgressBarWorker;
 
     /**
      * @var int Current Worker state
@@ -21,11 +18,6 @@ abstract class ParallelWorker implements Contracts\ParallelWorker {
      * @var string Worker Identifier
      */
     private string $identifier;
-
-    /**
-     * @var Channel|Closure|null Channel of communication between Task and ProgressBar
-     */
-    private Channel | Closure | null $progressBarChannel = null;
 
     /**
      * @var float Time when process started
@@ -44,21 +36,6 @@ abstract class ParallelWorker implements Contracts\ParallelWorker {
 
     final public function getState(): int {
         return $this->state;
-    }
-
-    final public function connectProgressBar(string | Closure $uuid, string $identifier = null): bool {
-        if ( !PARALLEL_EXT_LOADED) {
-            $this->progressBarChannel = $uuid;
-
-            return true;
-        }
-
-        // store worker identifier
-        $this->identifier = $identifier;
-        // connect to channel
-        $this->progressBarChannel = Channel::open(sprintf('progress-bar@%s', $uuid));
-
-        return true;
     }
 
     final public function start(...$args): void {
@@ -83,26 +60,6 @@ abstract class ParallelWorker implements Contracts\ParallelWorker {
      */
     abstract protected function process(): mixed;
 
-    final public function setMessage(string $message, string $name = 'message'): void {
-        $this->newProgressBarAction(__FUNCTION__, $message, $name);
-    }
-
-    final public function advance(int $steps = 1): void {
-        $this->newProgressBarAction(__FUNCTION__, $steps);
-    }
-
-    final public function setProgress(int $step): void {
-        $this->newProgressBarAction(__FUNCTION__, $step);
-    }
-
-    final public function display(): void {
-        $this->newProgressBarAction(__FUNCTION__);
-    }
-
-    final public function clear(): void {
-        $this->newProgressBarAction(__FUNCTION__);
-    }
-
     final public function getStartedAt(): ?float {
         return $this->started_at ?? null;
     }
@@ -117,28 +74,6 @@ abstract class ParallelWorker implements Contracts\ParallelWorker {
         }
 
         return $this->result;
-    }
-
-    private function newProgressBarAction(string $action, ...$args): void {
-        return;
-        // check if parallel is available
-        if (PARALLEL_EXT_LOADED) {
-            // report memory usage
-            $this->progressBarChannel->send(new StatsReportMessage(
-                worker_id:    $this->identifier,
-                memory_usage: memory_get_usage(),
-            ));
-            // request ProgressBar action
-            $this->progressBarChannel->send(new ProgressBarActionMessage(
-                action: $action,
-                args:   $args,
-            ));
-
-            return;
-        }
-
-        // redirect action to ProgressBar executor
-        ($this->progressBarChannel)($action, $args);
     }
 
 }
